create packages for movies and tvshows
