+++ app/src/main/java/mobile/kamheisiu/usmovientv/adapter/MoviesFragmentAdapter.java
@@ -4,17 +4,14 @@ import android.content.Context;
 import android.databinding.DataBindingUtil;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
-import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ImageView;
-import android.widget.TextView;
 
 import java.util.List;
 
 import mobile.kamheisiu.usmovientv.R;
 import mobile.kamheisiu.usmovientv.data.model.Movie;
 import mobile.kamheisiu.usmovientv.databinding.MovieItemBinding;
-import mobile.kamheisiu.usmovientv.viewmodel.ItemMovieViewModel;
+import mobile.kamheisiu.usmovientv.viewmodel.movies.ItemMovieViewModel;
 
 /**
  * Created by kamheisiu on 12/11/2017.
