+++ app/src/main/java/mobile/kamheisiu/usmovientv/fragment/movies/MoviesFragment.java
@@ -1,4 +1,4 @@
-package mobile.kamheisiu.usmovientv.fragment;
+package mobile.kamheisiu.usmovientv.fragment.movies;
 
 import android.databinding.DataBindingUtil;
 import android.os.Bundle;
@@ -17,26 +17,24 @@ import io.reactivex.functions.Consumer;
 import io.reactivex.schedulers.Schedulers;
 import mobile.kamheisiu.usmovientv.R;
 import mobile.kamheisiu.usmovientv.adapter.MoviesFragmentAdapter;
-import mobile.kamheisiu.usmovientv.adapter.TVShowsFragmentAdapter;
 import mobile.kamheisiu.usmovientv.data.model.Movie;
-import mobile.kamheisiu.usmovientv.data.model.TVShow;
-import mobile.kamheisiu.usmovientv.data.remote.TVShowsRequestResponse;
-import mobile.kamheisiu.usmovientv.databinding.FragmentTvShowsBinding;
-import mobile.kamheisiu.usmovientv.viewmodel.TVShowsFragmentViewModel;
+import mobile.kamheisiu.usmovientv.data.remote.MoviesRequestResponse;
+import mobile.kamheisiu.usmovientv.databinding.FragmentMoviesBinding;
+import mobile.kamheisiu.usmovientv.viewmodel.movies.MoviesFragmentViewModel;
 
 /**
  * Created by kamheisiu on 11/11/2017.
  */
 
-public abstract class TVShowsFragment extends Fragment {
+public abstract class MoviesFragment extends Fragment {
 
-    public static final String TAG = "TVShowsFragment";
+    public static final String TAG = "MoviesFragment";
     public static final String TITLE_KEY = "TITLE_KEY";
 
-    private FragmentTvShowsBinding binding;
+    private FragmentMoviesBinding binding;
     private String title;
 
-    protected TVShowsFragmentViewModel mTVShowsFragmentViewModel;
+    protected MoviesFragmentViewModel mMoviesFragmentViewModel;
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
@@ -51,7 +49,7 @@ public abstract class TVShowsFragment extends Fragment {
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_tv_shows, container, false);
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_movies, container, false);
 
         binding.swipeRefreshLayout.setOnRefreshListener(() -> { onRefresh(); });
 
@@ -66,7 +64,7 @@ public abstract class TVShowsFragment extends Fragment {
     public void onResume() {
         super.onResume();
 
-        if (mTVShowsFragmentViewModel.getCurrentTVShowsList() == null) {
+        if (mMoviesFragmentViewModel.getCurrentGetMoviesList() == null) {
             showSpinnerOnly();
             subscribeToGetMoviesResponse(true);
         }
@@ -76,7 +74,7 @@ public abstract class TVShowsFragment extends Fragment {
     public void onPause() {
         super.onPause();
 
-        mTVShowsFragmentViewModel.getTVShowsRequestResponse().unsubscribeOn(Schedulers.io());
+        mMoviesFragmentViewModel.getGetMoviesResponse().unsubscribeOn(Schedulers.io());
     }
 
     private void showHideComponents(boolean ifSpinnerVisible, boolean ifErrorMsgVisible, boolean ifRecyclerViewVisible) {
@@ -90,27 +88,27 @@ public abstract class TVShowsFragment extends Fragment {
     }
 
     private void subscribeToGetMoviesResponse(boolean isCreateNewVM) {
-        Consumer<TVShowsRequestResponse> onNext = tvShowsRequestResponse -> { onReceiveResponse(tvShowsRequestResponse); };
+        Consumer<MoviesRequestResponse> onNext = moviesRequestResponse -> { onReceiveResponse(moviesRequestResponse); };
 
         if (isCreateNewVM) {
             initViewModel();
         }
 
-        mTVShowsFragmentViewModel.getTVShowsRequestResponse()
+        mMoviesFragmentViewModel.getGetMoviesResponse()
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribe(onNext);
     }
 
-    private void onReceiveResponse(TVShowsRequestResponse response) {
+    private void onReceiveResponse(MoviesRequestResponse response) {
         binding.spinnerLoader.setVisibility(View.INVISIBLE);
         binding.swipeRefreshLayout.setEnabled(true);
 
         if (response.isSuccessful()) {
-            showTVShowsList(response.getGetTVShowsList().getTVShows());
+            showMoviesList(response.getGetMoviesList().getMovies());
         } else {
 
-            if (response.isRefresh() && mTVShowsFragmentViewModel.getCurrentTVShowsList() != null) {
+            if (response.isRefresh() && mMoviesFragmentViewModel.getCurrentGetMoviesList() != null) {
                 // we are already displaying a list
                 showHideComponents(false, false, true);
                 Toast.makeText(this.getContext(), getString(R.string.restful_call_try_again), Toast.LENGTH_SHORT).show();
@@ -120,22 +118,22 @@ public abstract class TVShowsFragment extends Fragment {
         }
     }
 
-    private void showTVShowsList(List<TVShow> tvShows) {
-        TVShowsFragmentAdapter adapter = new TVShowsFragmentAdapter(getContext(), tvShows);
+    private void showMoviesList(List<Movie> movies) {
+        MoviesFragmentAdapter adapter = new MoviesFragmentAdapter(this.getContext(), movies);
         binding.recyclerView.setAdapter(adapter);
 
-        GridLayoutManager gridLayoutManager = new GridLayoutManager(getContext(), 3);
+        GridLayoutManager gridLayoutManager = new GridLayoutManager(this.getContext(), 3);
         binding.recyclerView.setLayoutManager(gridLayoutManager);
 
         showHideComponents(false, false, true);
     }
 
     private void onRefresh() {
-        mTVShowsFragmentViewModel.getTVShowsRequestResponse().unsubscribeOn(Schedulers.io());
+        mMoviesFragmentViewModel.getGetMoviesResponse().unsubscribeOn(Schedulers.io());
 
         binding.swipeRefreshLayout.setRefreshing(false);
         binding.swipeRefreshLayout.setEnabled(false);
-        mTVShowsFragmentViewModel.onRefresh();
+        mMoviesFragmentViewModel.onRefresh();
 
         showSpinnerOnly();
         subscribeToGetMoviesResponse(false);
