+++ app/src/main/java/mobile/kamheisiu/usmovientv/fragment/movies/PopularMoviesFragment.java
@@ -1,8 +1,8 @@
-package mobile.kamheisiu.usmovientv.fragment;
+package mobile.kamheisiu.usmovientv.fragment.movies;
 
 import android.os.Bundle;
 
-import mobile.kamheisiu.usmovientv.viewmodel.PopularMoviesFragmentViewModel;
+import mobile.kamheisiu.usmovientv.viewmodel.movies.PopularMoviesFragmentViewModel;
 
 /**
  * Created by kamheisiu on 23/11/2017.
