+++ app/src/main/java/mobile/kamheisiu/usmovientv/fragment/movies/TopRatedMoviesFragment.java
@@ -1,9 +1,8 @@
-package mobile.kamheisiu.usmovientv.fragment;
+package mobile.kamheisiu.usmovientv.fragment.movies;
 
 import android.os.Bundle;
 
-import mobile.kamheisiu.usmovientv.viewmodel.NowPlayingMoviesFragmentViewModel;
-import mobile.kamheisiu.usmovientv.viewmodel.TopRatedMoviesFragmentViewModel;
+import mobile.kamheisiu.usmovientv.viewmodel.movies.TopRatedMoviesFragmentViewModel;
 
 /**
  * Created by kamheisiu on 23/11/2017.
