+++ app/src/main/java/mobile/kamheisiu/usmovientv/fragment/tvshows/OnTheAirTVShowsFragment.java
@@ -1,9 +1,8 @@
-package mobile.kamheisiu.usmovientv.fragment;
+package mobile.kamheisiu.usmovientv.fragment.tvshows;
 
 import android.os.Bundle;
 
-import mobile.kamheisiu.usmovientv.viewmodel.AiringTodayTVShowsFragmentViewModel;
-import mobile.kamheisiu.usmovientv.viewmodel.OnTheAirTVShowsFragmentViewModel;
+import mobile.kamheisiu.usmovientv.viewmodel.tvshows.OnTheAirTVShowsFragmentViewModel;
 
 /**
  * Created by kamheisiu on 23/11/2017.
