+++ app/src/main/java/mobile/kamheisiu/usmovientv/fragment/tvshows/PopularTVShowsFragment.java
@@ -1,9 +1,8 @@
-package mobile.kamheisiu.usmovientv.fragment;
+package mobile.kamheisiu.usmovientv.fragment.tvshows;
 
 import android.os.Bundle;
 
-import mobile.kamheisiu.usmovientv.viewmodel.AiringTodayTVShowsFragmentViewModel;
-import mobile.kamheisiu.usmovientv.viewmodel.PopularTVShowsFragmentViewModel;
+import mobile.kamheisiu.usmovientv.viewmodel.tvshows.PopularTVShowsFragmentViewModel;
 
 /**
  * Created by kamheisiu on 23/11/2017.
