+++ app/src/main/java/mobile/kamheisiu/usmovientv/viewmodel/movies/MovieDetailsActivityViewModel.java
@@ -1,8 +1,9 @@
-package mobile.kamheisiu.usmovientv.viewmodel;
+package mobile.kamheisiu.usmovientv.viewmodel.movies;
 
 import android.content.Context;
 import android.databinding.BaseObservable;
 import android.databinding.BindingAdapter;
+import android.util.Log;
 import android.widget.ImageView;
 
 import com.bumptech.glide.Glide;
@@ -13,46 +14,50 @@ import java.util.List;
 import io.reactivex.Observer;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.disposables.Disposable;
+import io.reactivex.functions.Consumer;
 import io.reactivex.schedulers.Schedulers;
 import io.reactivex.subjects.PublishSubject;
 import mobile.kamheisiu.usmovientv.R;
-import mobile.kamheisiu.usmovientv.data.model.CreatedBy;
+import mobile.kamheisiu.usmovientv.activity.MovieDetailsActivity;
 import mobile.kamheisiu.usmovientv.data.model.Genre;
+import mobile.kamheisiu.usmovientv.data.model.Movie;
+import mobile.kamheisiu.usmovientv.data.model.MovieDetails;
+import mobile.kamheisiu.usmovientv.data.model.ProductionCompany;
+import mobile.kamheisiu.usmovientv.data.model.ProductionCountry;
 import mobile.kamheisiu.usmovientv.data.model.SpokenLanguage;
-import mobile.kamheisiu.usmovientv.data.model.TVShowDetails;
 import mobile.kamheisiu.usmovientv.data.remote.ApiUtils;
-import mobile.kamheisiu.usmovientv.data.remote.TVShowsServices;
+import mobile.kamheisiu.usmovientv.data.remote.MoviesServices;
 
 /**
- * Created by kamheisiu on 24/11/2017.
+ * Created by kamheisiu on 21/11/2017.
  */
 
-public class TVShowDetailsActivityViewModel extends BaseObservable {
+public class MovieDetailsActivityViewModel extends BaseObservable {
 
     private Context mContext;
-    private int mTvShowId;
-    private TVShowDetails mTVShowDetails;
+    private int mMovieId;
+    private MovieDetails mMovieDetails;
     private PublishSubject<Boolean> isDataReady = PublishSubject.create();
 
-    public TVShowDetailsActivityViewModel(Context context, int tvShowId) {
+    public MovieDetailsActivityViewModel(Context context, int movieId) {
         mContext = context;
-        mTvShowId = tvShowId;
-        getTVShowDetails();
+        mMovieId = movieId;
+        getMovieDetails();
     }
 
-    private void getTVShowDetails() {
-        TVShowsServices tvShowsServices = new ApiUtils().getTVShowsServices();
+    private void getMovieDetails() {
+        MoviesServices moviesServices = new ApiUtils().getMoviesServices();
 
-        tvShowsServices.getTVShowDetails(mTvShowId).subscribeOn(Schedulers.io()).observeOn(AndroidSchedulers.mainThread())
-                .subscribe(new Observer<TVShowDetails>() {
+        moviesServices.getMovieDetails(mMovieId).subscribeOn(Schedulers.io()).observeOn(AndroidSchedulers.mainThread())
+                .subscribe(new Observer<MovieDetails>() {
                     @Override
                     public void onSubscribe(Disposable d) {
 
                     }
 
                     @Override
-                    public void onNext(TVShowDetails tvShowDetails) {
-                        mTVShowDetails = tvShowDetails;
+                    public void onNext(MovieDetails movieDetails) {
+                        mMovieDetails = movieDetails;
                         isDataReady.onNext(true);
                     }
 
@@ -68,26 +73,30 @@ public class TVShowDetailsActivityViewModel extends BaseObservable {
                 });
     }
 
+    public PublishSubject<Boolean> getIsDataReady() {
+        return isDataReady;
+    }
+
     public String getPopularity() {
-        return String.format("☆ %.1f", mTVShowDetails.getPopularity());
+        return String.format("☆ %.1f", mMovieDetails.getPopularity());
     }
 
     public String getVoteAverage() {
-        return String.format("♡ %.1f", mTVShowDetails.getVoteAverage());
+        return String.format("♡ %.1f", mMovieDetails.getVoteAverage());
     }
 
     public String getVoteCount() {
-        return String.format("웃 %d", mTVShowDetails.getVoteCount());
+        return String.format("웃 %d", mMovieDetails.getVoteCount());
     }
 
     public String getTitle() {
-        return mTVShowDetails.getName();
+        return mMovieDetails.getTitle();
     }
 
     public String getGenres() {
         StringBuilder genresBuilder = new StringBuilder();
 
-        List<Genre> genres = mTVShowDetails.getGenres();
+        List<Genre> genres = mMovieDetails.getGenres();
         for (int i=0; i<genres.size(); i++) {
             if (i == genres.size() - 1) {
                 genresBuilder.append(genres.get(i).getName());
@@ -102,66 +111,58 @@ public class TVShowDetailsActivityViewModel extends BaseObservable {
     public String getLanguages() {
         StringBuilder languagesBuilder = new StringBuilder();
 
-        List<String> languages = mTVShowDetails.getLanguages();
+        List<SpokenLanguage> languages = mMovieDetails.getSpokenLanguages();
         for (int i=0; i<languages.size(); i++) {
             if (i == languages.size() - 1) {
-                languagesBuilder.append(languages.get(i));
+                languagesBuilder.append(languages.get(i).getName());
             } else {
-                languagesBuilder.append(languages.get(i)).append(", ");
+                languagesBuilder.append(languages.get(i).getName()).append(", ");
             }
         }
 
         return languagesBuilder.toString();
     }
 
-    public String getCreatedBy() {
-        StringBuilder createdByListBuilder = new StringBuilder();
+    public String getCompanies() {
+        StringBuilder companiesBuilder = new StringBuilder();
 
-        List<CreatedBy> createdBy = mTVShowDetails.getCreatedBy();
-        for (int i=0; i<createdBy.size(); i++) {
-            if (i == createdBy.size() - 1) {
-                createdByListBuilder.append(createdBy.get(i).getName());
+        List<ProductionCompany> companies = mMovieDetails.getProductionCompanies();
+        for (int i=0; i<companies.size(); i++) {
+            if (i == companies.size() - 1) {
+                companiesBuilder.append(companies.get(i).getName());
             } else {
-                createdByListBuilder.append(createdBy.get(i).getName()).append(", ");
+                companiesBuilder.append(companies.get(i).getName()).append(", ");
             }
         }
 
-        return createdByListBuilder.toString();
+        return companiesBuilder.toString();
     }
 
-    public String getNumberOfSeasons() {
-        return mTVShowDetails.getNumberOfSeasons() + "";
+    public String getCountries() {
+        StringBuilder countriesBuilder = new StringBuilder();
+
+        List<ProductionCountry> countries = mMovieDetails.getProductionCountries();
+        for (int i=0; i<countries.size(); i++) {
+            if (i == countries.size() - 1) {
+                countriesBuilder.append(countries.get(i).getName());
+            } else {
+                countriesBuilder.append(countries.get(i).getName()).append(", ");
+            }
+        }
+
+        return countriesBuilder.toString();
     }
 
-    public String getNumberOfEpisodes() {
-        return mTVShowDetails.getNumberOfEpisodes() + "";
+    public String getReleaseDate() {
+        return mMovieDetails.getReleaseDate();
     }
 
     public String getRuntime() {
-        StringBuilder runtimeBuilder = new StringBuilder();
-
-        List<Integer> runtimes = mTVShowDetails.getEpisodeRunTime();
-        for (int i=0; i<runtimes.size(); i++) {
-            if (i == runtimes.size() - 1) {
-                runtimeBuilder.append(runtimes.get(i));
-            } else {
-                runtimeBuilder.append(runtimes.get(i)).append(", ");
-            }
-        }
-
-        return runtimeBuilder.toString();
-    }
-
-    public String getOverview() {
-        return mTVShowDetails.getOverview();
-    }
-
-    public PublishSubject<Boolean> getIsDataReady() {
-        return isDataReady;
+        return mMovieDetails.getRuntime() + "";
     }
 
     public String getPosterPath() {
-        return mTVShowDetails.getPosterPath();
+        return mMovieDetails.getPosterPath();
     }
 
     @BindingAdapter({"image"})
