+++ app/src/main/java/mobile/kamheisiu/usmovientv/viewmodel/tvshows/ItemTVShowViewModel.java
@@ -1,10 +1,9 @@
-package mobile.kamheisiu.usmovientv.viewmodel;
+package mobile.kamheisiu.usmovientv.viewmodel.tvshows;
 
 import android.content.Context;
 import android.content.Intent;
 import android.databinding.BaseObservable;
 import android.databinding.BindingAdapter;
-import android.net.Uri;
 import android.util.Log;
 import android.view.View;
 import android.widget.ImageView;
@@ -14,22 +13,22 @@ import com.bumptech.glide.request.RequestOptions;
 
 import mobile.kamheisiu.usmovientv.R;
 import mobile.kamheisiu.usmovientv.activity.MovieDetailsActivity;
+import mobile.kamheisiu.usmovientv.activity.TVShowDetailsActivity;
 import mobile.kamheisiu.usmovientv.data.model.Movie;
+import mobile.kamheisiu.usmovientv.data.model.TVShow;
 import mobile.kamheisiu.usmovientv.data.remote.ApiUtils;
 
-import static mobile.kamheisiu.usmovientv.data.remote.ApiUtils.IMAGE_BASE_URL;
-
 /**
- * Created by kamheisiu on 12/11/2017.
+ * Created by kamheisiu on 23/11/2017.
  */
 
-public class ItemMovieViewModel extends BaseObservable {
+public class ItemTVShowViewModel extends BaseObservable {
 
-    private Movie mMovie;
+    private TVShow mTVShow;
     private Context mContext;
 
-    public ItemMovieViewModel(Movie movie, Context context) {
-        mMovie = movie;
+    public ItemTVShowViewModel(TVShow tvShow, Context context) {
+        mTVShow = tvShow;
         mContext = context;
     }
 
@@ -48,30 +47,30 @@ public class ItemMovieViewModel extends BaseObservable {
         Glide.with(view.getContext())
                 .load(urlBuilder.toString())
                 .apply(new RequestOptions()
-                    .placeholder(R.drawable.movietv_clear_bg)
-                    .centerCrop()
-                    .dontAnimate()
-                    .dontTransform())
+                        .placeholder(R.drawable.movietv_clear_bg)
+                        .centerCrop()
+                        .dontAnimate()
+                        .dontTransform())
                 .into(view);
     }
 
     public View.OnClickListener onItemClick() {
         return v -> {
-            Intent intent = new Intent(mContext, MovieDetailsActivity.class);
-            intent.putExtra(MovieDetailsActivity.MOVIE_ID_KEY, mMovie.getId());
+            Intent intent = new Intent(mContext, TVShowDetailsActivity.class);
+            intent.putExtra(TVShowDetailsActivity.TV_SHOW_ID_KEY, mTVShow.getId());
             mContext.startActivity(intent);
         };
     }
 
     public String getReleaseDate() {
-        return mMovie.getReleaseDate();
+        return mTVShow.getFirstAirDate();
     }
 
     public String getPosterUrl() {
-        return mMovie.getPosterPath();
+        return mTVShow.getPosterPath();
     }
 
     public String getPopularity() {
-        return String.format("%.2f", mMovie.getPopularity());
+        return String.format("%.2f", mTVShow.getPopularity());
     }
 }
