+++ app/src/main/java/mobile/kamheisiu/usmovientv/viewmodel/tvshows/TVShowDetailsActivityViewModel.java
@@ -1,9 +1,8 @@
-package mobile.kamheisiu.usmovientv.viewmodel;
+package mobile.kamheisiu.usmovientv.viewmodel.tvshows;
 
 import android.content.Context;
 import android.databinding.BaseObservable;
 import android.databinding.BindingAdapter;
-import android.util.Log;
 import android.widget.ImageView;
 
 import com.bumptech.glide.Glide;
@@ -14,50 +13,46 @@ import java.util.List;
 import io.reactivex.Observer;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.disposables.Disposable;
-import io.reactivex.functions.Consumer;
 import io.reactivex.schedulers.Schedulers;
 import io.reactivex.subjects.PublishSubject;
 import mobile.kamheisiu.usmovientv.R;
-import mobile.kamheisiu.usmovientv.activity.MovieDetailsActivity;
+import mobile.kamheisiu.usmovientv.data.model.CreatedBy;
 import mobile.kamheisiu.usmovientv.data.model.Genre;
-import mobile.kamheisiu.usmovientv.data.model.Movie;
-import mobile.kamheisiu.usmovientv.data.model.MovieDetails;
-import mobile.kamheisiu.usmovientv.data.model.ProductionCompany;
-import mobile.kamheisiu.usmovientv.data.model.ProductionCountry;
 import mobile.kamheisiu.usmovientv.data.model.SpokenLanguage;
+import mobile.kamheisiu.usmovientv.data.model.TVShowDetails;
 import mobile.kamheisiu.usmovientv.data.remote.ApiUtils;
-import mobile.kamheisiu.usmovientv.data.remote.MoviesServices;
+import mobile.kamheisiu.usmovientv.data.remote.TVShowsServices;
 
 /**
- * Created by kamheisiu on 21/11/2017.
+ * Created by kamheisiu on 24/11/2017.
  */
 
-public class MovieDetailsActivityViewModel extends BaseObservable {
+public class TVShowDetailsActivityViewModel extends BaseObservable {
 
     private Context mContext;
-    private int mMovieId;
-    private MovieDetails mMovieDetails;
+    private int mTvShowId;
+    private TVShowDetails mTVShowDetails;
     private PublishSubject<Boolean> isDataReady = PublishSubject.create();
 
-    public MovieDetailsActivityViewModel(Context context, int movieId) {
+    public TVShowDetailsActivityViewModel(Context context, int tvShowId) {
         mContext = context;
-        mMovieId = movieId;
-        getMovieDetails();
+        mTvShowId = tvShowId;
+        getTVShowDetails();
     }
 
-    private void getMovieDetails() {
-        MoviesServices moviesServices = new ApiUtils().getMoviesServices();
+    private void getTVShowDetails() {
+        TVShowsServices tvShowsServices = new ApiUtils().getTVShowsServices();
 
-        moviesServices.getMovieDetails(mMovieId).subscribeOn(Schedulers.io()).observeOn(AndroidSchedulers.mainThread())
-                .subscribe(new Observer<MovieDetails>() {
+        tvShowsServices.getTVShowDetails(mTvShowId).subscribeOn(Schedulers.io()).observeOn(AndroidSchedulers.mainThread())
+                .subscribe(new Observer<TVShowDetails>() {
                     @Override
                     public void onSubscribe(Disposable d) {
 
                     }
 
                     @Override
-                    public void onNext(MovieDetails movieDetails) {
-                        mMovieDetails = movieDetails;
+                    public void onNext(TVShowDetails tvShowDetails) {
+                        mTVShowDetails = tvShowDetails;
                         isDataReady.onNext(true);
                     }
 
@@ -73,30 +68,26 @@ public class MovieDetailsActivityViewModel extends BaseObservable {
                 });
     }
 
-    public PublishSubject<Boolean> getIsDataReady() {
-        return isDataReady;
-    }
-
     public String getPopularity() {
-        return String.format("☆ %.1f", mMovieDetails.getPopularity());
+        return String.format("☆ %.1f", mTVShowDetails.getPopularity());
     }
 
     public String getVoteAverage() {
-        return String.format("♡ %.1f", mMovieDetails.getVoteAverage());
+        return String.format("♡ %.1f", mTVShowDetails.getVoteAverage());
     }
 
     public String getVoteCount() {
-        return String.format("웃 %d", mMovieDetails.getVoteCount());
+        return String.format("웃 %d", mTVShowDetails.getVoteCount());
     }
 
     public String getTitle() {
-        return mMovieDetails.getTitle();
+        return mTVShowDetails.getName();
     }
 
     public String getGenres() {
         StringBuilder genresBuilder = new StringBuilder();
 
-        List<Genre> genres = mMovieDetails.getGenres();
+        List<Genre> genres = mTVShowDetails.getGenres();
         for (int i=0; i<genres.size(); i++) {
             if (i == genres.size() - 1) {
                 genresBuilder.append(genres.get(i).getName());
@@ -111,58 +102,66 @@ public class MovieDetailsActivityViewModel extends BaseObservable {
     public String getLanguages() {
         StringBuilder languagesBuilder = new StringBuilder();
 
-        List<SpokenLanguage> languages = mMovieDetails.getSpokenLanguages();
+        List<String> languages = mTVShowDetails.getLanguages();
         for (int i=0; i<languages.size(); i++) {
             if (i == languages.size() - 1) {
-                languagesBuilder.append(languages.get(i).getName());
+                languagesBuilder.append(languages.get(i));
             } else {
-                languagesBuilder.append(languages.get(i).getName()).append(", ");
+                languagesBuilder.append(languages.get(i)).append(", ");
             }
         }
 
         return languagesBuilder.toString();
     }
 
-    public String getCompanies() {
-        StringBuilder companiesBuilder = new StringBuilder();
+    public String getCreatedBy() {
+        StringBuilder createdByListBuilder = new StringBuilder();
 
-        List<ProductionCompany> companies = mMovieDetails.getProductionCompanies();
-        for (int i=0; i<companies.size(); i++) {
-            if (i == companies.size() - 1) {
-                companiesBuilder.append(companies.get(i).getName());
+        List<CreatedBy> createdBy = mTVShowDetails.getCreatedBy();
+        for (int i=0; i<createdBy.size(); i++) {
+            if (i == createdBy.size() - 1) {
+                createdByListBuilder.append(createdBy.get(i).getName());
             } else {
-                companiesBuilder.append(companies.get(i).getName()).append(", ");
+                createdByListBuilder.append(createdBy.get(i).getName()).append(", ");
             }
         }
 
-        return companiesBuilder.toString();
+        return createdByListBuilder.toString();
     }
 
-    public String getCountries() {
-        StringBuilder countriesBuilder = new StringBuilder();
-
-        List<ProductionCountry> countries = mMovieDetails.getProductionCountries();
-        for (int i=0; i<countries.size(); i++) {
-            if (i == countries.size() - 1) {
-                countriesBuilder.append(countries.get(i).getName());
-            } else {
-                countriesBuilder.append(countries.get(i).getName()).append(", ");
-            }
-        }
-
-        return countriesBuilder.toString();
+    public String getNumberOfSeasons() {
+        return mTVShowDetails.getNumberOfSeasons() + "";
     }
 
-    public String getReleaseDate() {
-        return mMovieDetails.getReleaseDate();
+    public String getNumberOfEpisodes() {
+        return mTVShowDetails.getNumberOfEpisodes() + "";
     }
 
     public String getRuntime() {
-        return mMovieDetails.getRuntime() + "";
+        StringBuilder runtimeBuilder = new StringBuilder();
+
+        List<Integer> runtimes = mTVShowDetails.getEpisodeRunTime();
+        for (int i=0; i<runtimes.size(); i++) {
+            if (i == runtimes.size() - 1) {
+                runtimeBuilder.append(runtimes.get(i));
+            } else {
+                runtimeBuilder.append(runtimes.get(i)).append(", ");
+            }
+        }
+
+        return runtimeBuilder.toString();
+    }
+
+    public String getOverview() {
+        return mTVShowDetails.getOverview();
+    }
+
+    public PublishSubject<Boolean> getIsDataReady() {
+        return isDataReady;
     }
 
     public String getPosterPath() {
-        return mMovieDetails.getPosterPath();
+        return mTVShowDetails.getPosterPath();
     }
 
     @BindingAdapter({"image"})
