+++ app/src/main/res/layout/activity_movie_details.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="mdavm"
-            type="mobile.kamheisiu.usmovientv.viewmodel.MovieDetailsActivityViewModel"/>
+            type="mobile.kamheisiu.usmovientv.viewmodel.movies.MovieDetailsActivityViewModel"/>
     </data>
 
     <RelativeLayout
