+++ app/src/main/res/layout/movie_item.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="movievm"
-            type="mobile.kamheisiu.usmovientv.viewmodel.ItemMovieViewModel"/>
+            type="mobile.kamheisiu.usmovientv.viewmodel.movies.ItemMovieViewModel"/>
     </data>
 
     <RelativeLayout
